bug fix homework1

+++ js/task-3.js
@@ -1,7 +1,7 @@
 function getElementWidth(content, padding, border) {
   return (
-    Number.parseInt(content) +
-    (Number.parseInt(padding) + Number.parseFloat(border)) * 2
+    Number.parseFloat(content) +
+    (Number.parseFloat(padding) + Number.parseFloat(border)) * 2
   );
 }
